index_pipeline/requirements.txt
# Core LlamaIndex components
llama-index-core>=0.10.0
llama-index-embeddings-huggingface>=0.2.0
llama-index-vector-stores-chroma>=0.1.0

# Vector database
chromadb>=0.4.0

# Document processing
PyPDF2>=3.0.0
python-pptx>=0.6.21
python-docx>=0.8.11

# Image processing and OCR
Pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.16.0

# Embedding models (HuggingFace)
sentence-transformers>=2.2.0
torch>=2.0.0

# Utilities
pathlib>=1.0.1
uuid

# Optional: For better performance
accelerate>=0.20.0